move getValue logic to property
